refactoring index.js for Standards folder
