+++ components/LessonSection/Preview/LessonsCarousel.js
@@ -27,12 +27,14 @@ import { customControls, getVideoThumb } from './utils';
 import { Button } from 'react-bootstrap';
 import { FaArrowLeft, FaArrowRight } from 'react-icons/fa';
 import Image from 'next/image';
+import useLessonElementInView from '../../../customHooks/useLessonElementInView';
 
 
-const LessonsCarousel = ({ mediaItems }) => {
+const LessonsCarousel = ({ mediaItems, _sectionDots, SectionTitle }) => {
     const [currentIndex, setCurrentIndex] = useState(0);
     const mediaItemsSorted = mediaItems.sort((lessonDocumentA, lessonDocumentB) => lessonDocumentA.order - lessonDocumentB.order).map((item, index) => ({ ...item, isVisible: index === 0 }));
     const [controlDots, setControlDots] = useState(mediaItemsSorted);
+    const { ref } = useLessonElementInView(_sectionDots, SectionTitle)
 
     const handleNextBtnClick = () => {
         console.log('next btn was clicked: ');
@@ -78,7 +80,7 @@ const LessonsCarousel = ({ mediaItems }) => {
 
 
     return (
-        <div className={`bg-light-gray rounded p-sm-3 display-flex carouselSelectedLessons flex-column justify-content-center align-items-center ${styles.Carousel}`} >
+        <div ref={ref} className={`bg-light-gray rounded p-sm-3 display-flex carouselSelectedLessons flex-column justify-content-center align-items-center ${styles.Carousel}`} >
             <section style={{ height: 'fit-content' }} className="autoCarouselContainer">
                 <div className="autoCarouselSlider" style={{ transform: `translate3d(${-currentIndex * 100}%, 0, 0)` }}>
                     {mediaItems && mediaItems.sort((lessonDocumentA, lessonDocumentB) => lessonDocumentA.order - lessonDocumentB.order).map((lessonDocument, index) => <LessonSlide key={index} {...lessonDocument} />)}
@@ -135,11 +137,11 @@ const LessonsCarousel = ({ mediaItems }) => {
                                             fill
                                         />
                                     </div>
-                                :
-                                <i
-                                    key={index}
-                                    className="bi-filetype-pdf fs-2"
-                                />}
+                                    :
+                                    <i
+                                        key={index}
+                                        className="bi-filetype-pdf fs-2"
+                                    />}
                             </li>
                         )
                     })}
