+++ components/LessonSection/Standards/StandardsGroup.js
@@ -19,7 +19,7 @@ export const formatAlignmentNotes = (text) => {
   return text.replace(/•/g, '-').replace(/\^2/g, '²');
 };
 
-const StandardsGroup = ({ 
+const StandardsGroup = ({
   id,
   codes,
   grades,
