+++ components/LessonSection/TeachIt/index.js
@@ -7,6 +7,7 @@ import CollapsibleLessonSection from '../../CollapsibleLessonSection';
 import LessonPart from './LessonPart';
 import { ModalContext } from '../../../providers/ModalProvider';
 import { useContext } from 'react';
+import useLessonElementInView from '../../../customHooks/useLessonElementInView';
 
 const getIsValObj = val => (typeof val === 'object') && !Array.isArray(val) && (val !== null);
 const getObjVals = obj => {
@@ -17,7 +18,7 @@ const getObjVals = obj => {
     const val = obj[key];
     vals.push(val);
   });
-  
+
   return vals;
 };
 
@@ -36,9 +37,9 @@ const TeachIt = ({
   const [selectedEnvironment, setSelectedEnvironment] = useState(environments[0]);
   const allResources = getIsValObj(Data[selectedEnvironment].resources) ? getObjVals(Data[selectedEnvironment].resources) : Data[selectedEnvironment].resources;
   let resources = allResources.find(({ gradePrefix }) => gradePrefix === selectedGrade.gradePrefix);
-
   resources = getIsValObj(resources) ? [resources] : resources;
-  
+  const { ref } = useLessonElementInView(_sectionDots, SectionTitle);
+
   const handleIconClick = () => {
     console.log("clicked");
     setIsDownloadModalInfoOn(true);
@@ -57,7 +58,7 @@ const TeachIt = ({
       initiallyExpanded
       _sectionDots={_sectionDots}
     >
-      <>
+      <div ref={ref}>
         <div className='container row mx-auto mt-4'>
           <div className='col-12 col-xl-8 offset-xl-2 bg-light-gray p-3 row align-items-center gap-3 gap-lg-0'>
             <div className='fs-5 mb-2'>
@@ -141,7 +142,7 @@ const TeachIt = ({
             />
           ))}
         </div>
-      </>
+      </div>
     </CollapsibleLessonSection>
   );
 };
